requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pandas
matplotlib
pyttsx3
deep-translator
beautifulsoup4
requests
gtts
